requirements.txt
contourpy==1.3.1
cramjam==2.9.1
cycler==0.12.1
dnspython==2.7.0
faiss-cpu==1.10.0
fastparquet==2024.11.0
fonttools==4.56.0
fsspec==2025.3.0
joblib==1.4.2
kiwisolver==1.4.8
lightgbm>=3.3.0
matplotlib>=3.5.0
mistralai>=0.0.7
numpy>=1.24.0
packaging==24.2
pandas>=1.5.0
pillow==11.1.0
pymongo>=4.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv>=0.20.0
pytz==2025.1
scikit-learn>=1.0.0
scipy==1.15.2
seaborn>=0.12.0
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.1
transformers>=4.37.0
torch>=2.0.0
datasets>=2.14.0
